Тесты на сервисы поправил/

+++ src/main/java/ru/coffeecoders/questbot/entities/Question.java
@@ -39,7 +39,7 @@ public class Question {
         return questionId;
     }
 
-    public void setQuestionId(int questionId) {
+    public void setQuestionId(long questionId) {
         this.questionId = questionId;
     }
